IKASAN-1337 -  Refactor spring-boot-jms
