+++ ikasaneip/sample/spring-boot/jms/pom.xml
@@ -48,33 +48,45 @@
             <version>${version.org.springboot}</version>
         </dependency>
 
+        <!-- junit should be with scope test however for the sample purpose we
+        spin up the embedded broker as part of sample application.-->
+
         <dependency>
             <groupId>junit</groupId>
             <artifactId>junit</artifactId>
-            <scope>test</scope>
+            <!--<scope>test</scope>-->
         </dependency>
 
         <dependency>
-            <groupId>org.springframework.boot</groupId>
-            <artifactId>spring-boot-starter-activemq</artifactId>
-            <version>${version.org.springboot}</version>
+            <groupId>org.ikasan</groupId>
+            <artifactId>ikasan-test</artifactId>
             <scope>test</scope>
         </dependency>
+
 
         <dependency>
             <groupId>org.apache.activemq</groupId>
-            <artifactId>activemq-all</artifactId>
+            <artifactId>activemq-client</artifactId>
         </dependency>
+
+        <!-- activemq-all should be with scope test however for the sample purpose we
+         spin up the embedded broker as part of sample application.-->
+        <dependency>
+            <groupId>org.apache.activemq</groupId>
+            <artifactId>activemq-all</artifactId>
+            <!--<scope>test</scope>-->
+        </dependency>
+
+        <!-- activemq-junit should be with scope test however for the sample purpose we
+         spin up the embedded broker as part of sample application.-->
 
         <dependency>
             <groupId>org.apache.activemq.tooling</groupId>
             <artifactId>activemq-junit</artifactId>
-            <scope>test</scope>
+            <!--<scope>test</scope>-->
         </dependency>
 
-
     </dependencies>
-
 
     <build>
         <plugins>
