+++ ikasaneip/sample/spring-boot/jms/src/main/java/com/ikasan/sample/spring/boot/builderpattern/Application.java
@@ -1,5 +1,6 @@
 package com.ikasan.sample.spring.boot.builderpattern;
 
+import org.apache.activemq.junit.EmbeddedActiveMQBroker;
 import org.ikasan.builder.IkasanApplication;
 import org.ikasan.builder.IkasanApplicationFactory;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -12,7 +13,11 @@ public class Application
 
     public static void main(String[] args) throws Exception
     {
-        IkasanApplication ikasanApplication = IkasanApplicationFactory.getIkasanApplication(args);
+        // Embedded broker is started as part of sample
+        new EmbeddedActiveMQBroker().start();
+
+
+        IkasanApplicationFactory.getIkasanApplication(Application.class,args);
 
         System.out.println("Context ready");
     }
