+++ ikasaneip/sample/spring-boot/jms/src/main/resources/application.properties
@@ -1,4 +1,7 @@
-logging.level.com.arjuna=DEBUG
+logging.level.com.arjuna=INFO
+
+logging.level.com.ikasan=INFO
+logging.level.org.ikasan=INFO
 
 spring.jta.narayana.transaction-manager-id=2
 spring.jta.narayana.default-timeout=60
@@ -9,14 +12,13 @@ spring.jta.narayana.log-dir=./target/ObjectStore
 
 spring.mvc.view.prefix=/WEB-INF/jsp/
 spring.mvc.view.suffix=.jsp
-#security.ignored=/login*,/js/**,/css/**
 
-#server.error.whitelabel.enabled=false
+server.error.whitelabel.enabled=false
 
 
 server.port=8080
-service.name=localhost
-#server.address=0.0.0.0
+#service.name=localhost
+server.address=0.0.0.0
 server.contextPath=/sample-boot-jms
 
 management.security.enabled=false
@@ -32,11 +34,8 @@ datasource.dialect=org.hibernate.dialect.H2Dialect
 datasource.show-sql=false
 datasource.hbm2ddl.auto=create
 
-jms.provider.url=vm://embedded-broker?create=true
-jms.naming.factory.initial=org.apache.activemq.jndi.ActiveMQInitialContextFactory
-jms.naming.factory.url.pkgs=
-jms.naming.security.principal=
-jms.naming.security.credentials=
+jms.broker.name=embedded-broker
+jms.provider.url=failover:(vm://${jms.broker.name}?create=false)
 
 #jms.provider.url=remote://svc-messaging01:4447
 #jms.naming.factory.initial=org.jboss.naming.remote.client.InitialContextFactory
