--- ikasaneip/sample/spring-boot/jms/src/main/resources/log4j.xml
@@ -1,22 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE log4j:configuration SYSTEM "log4j.dtd">
-<log4j:configuration debug="false" xmlns:log4j='http://jakarta.apache.org/log4j/'>
-
-    <appender name="console" class="org.apache.log4j.ConsoleAppender">
-        <param name="Target" value="System.out"/>
-        <layout class="org.apache.log4j.PatternLayout">
-            <param name="ConversionPattern" value="%d{yyyyMMdd HH:mm:ss SSS} %-5p [%t]: %x %m%n"/>
-        </layout>
-    </appender>
-
-    <logger name="org.springframework.security.ldap">
-        <level value="WARN"/>
-        <appender-ref ref="console"/>
-    </logger>
-
-    <root>
-        <level value="INFO"/>
-        <appender-ref ref="console"/>
-    </root>
-
-</log4j:configuration>
