--- ikasaneip/sample/spring-boot/jms/src/main/resources/logger-conf.xml
@@ -1,18 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<beans xmlns="http://www.springframework.org/schema/beans"
-    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-    xsi:schemaLocation="
-        http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans-3.1.xsd">
-
-    <bean id="log4jInitializer" class="org.springframework.beans.factory.config.MethodInvokingFactoryBean">
-        <property name="staticMethod">
-            <value>org.springframework.util.Log4jConfigurer.initLogging</value>
-        </property>
-        <property name="arguments">
-            <list>
-                <value>classpath:log4j.xml</value>
-            </list>
-        </property>
-    </bean>
-
-</beans>
