+++ ikasaneip/sample/spring-boot/jms/src/test/java/com/ikasan/sample/spring/boot/builderpattern/ModuleTestConfig.java
@@ -1,13 +1,11 @@
 package com.ikasan.sample.spring.boot.builderpattern;
 
 import org.apache.activemq.ActiveMQConnectionFactory;
+import org.ikasan.testharness.flow.jms.MessageListenerVerifier;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.jms.annotation.EnableJms;
-import org.springframework.jms.config.JmsListenerContainerFactory;
 import org.springframework.jms.config.JmsListenerEndpointRegistry;
-import org.springframework.jms.config.SimpleJmsListenerContainerFactory;
 import org.springframework.jms.core.JmsTemplate;
 
 import javax.annotation.Resource;
