+++ ikasaneip/sample/spring-boot/jms/src/test/resources/application.properties
@@ -9,13 +9,12 @@ spring.jta.narayana.log-dir=./target/ObjectStore
 
 spring.mvc.view.prefix=/WEB-INF/jsp/
 spring.mvc.view.suffix=.jsp
-#security.ignored=/login*,/js/**,/css/**
 
-#server.error.whitelabel.enabled=false
+server.error.whitelabel.enabled=false
 
 server.port=8080
-service.name=sample-boot-jms
-#server.address=0.0.0.0
+#service.name=sample-boot-jms
+server.address=localhost
 server.contextPath=/sample-boot-jms
 
 management.security.enabled=false
@@ -34,15 +33,3 @@ datasource.hbm2ddl.auto=create
 
 jms.broker.name=embedded-broker
 jms.provider.url=failover:(vm://${jms.broker.name}?create=false)
-jms.naming.factory.initial=org.apache.activemq.jndi.ActiveMQInitialContextFactory
-jms.naming.factory.url.pkgs=
-jms.naming.security.principal=
-jms.naming.security.credentials=
-
-jms.source.destination=dynamicQueues/source
-jms.target.destination=dynamicQueues/target
-jms.connectionFactory.jndi.name=ConnectionFactory
-
-# Bootstrap activemq broker
-spring.activemq.broker-url=failover:(vm://embedded-broker?create=false)
-spring.activemq.inMemory=true
