+++ ikasaneip/test/pom.xml
@@ -1,30 +1,50 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<!-- # // # // # // $Id$ # // $URL$ # // # // ==================================================================== 
-	# // Ikasan Enterprise Integration Platform # // # // Distributed under the 
-	Modified BSD License. # // Copyright notice: The copyright for this software 
-	and a full listing # // of individual contributors are as shown in the packaged 
-	copyright.txt # // file. # // # // All rights reserved. # // # // Redistribution 
-	and use in source and binary forms, with or without # // modification, are 
-	permitted provided that the following conditions are met: # // # // - Redistributions 
-	of source code must retain the above copyright notice, # // this list of 
-	conditions and the following disclaimer. # // # // - Redistributions in binary 
-	form must reproduce the above copyright notice, # // this list of conditions 
-	and the following disclaimer in the documentation # // and/or other materials 
-	provided with the distribution. # // # // - Neither the name of the ORGANIZATION 
-	nor the names of its contributors may # // be used to endorse or promote 
-	products derived from this software without # // specific prior written permission. 
-	# // # // THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS 
-	"AS IS" # // AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED 
-	TO, THE # // IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR 
-	PURPOSE ARE # // DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS 
-	BE LIABLE # // FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, 
-	OR CONSEQUENTIAL # // DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT 
-	OF SUBSTITUTE GOODS OR # // SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS 
-	INTERRUPTION) HOWEVER # // CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER 
-	IN CONTRACT, STRICT LIABILITY, # // OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) 
-	ARISING IN ANY WAY OUT OF THE # // USE OF THIS SOFTWARE, EVEN IF ADVISED 
-	OF THE POSSIBILITY OF SUCH DAMAGE. # // ==================================================================== 
-	# // # // Author: Ikasan Development Team # // -->
+<!--
+
+
+ $Id$
+ $URL$
+
+ ====================================================================
+ Ikasan Enterprise Integration Platform
+
+ Distributed under the Modified BSD License.
+ Copyright notice: The copyright for this software and a full listing
+ of individual contributors are as shown in the packaged copyright.txt
+ file.
+
+ All rights reserved.
+
+ Redistribution and use in source and binary forms, with or without
+ modification, are permitted provided that the following conditions are met:
+
+  - Redistributions of source code must retain the above copyright notice,
+    this list of conditions and the following disclaimer.
+
+  - Redistributions in binary form must reproduce the above copyright notice,
+    this list of conditions and the following disclaimer in the documentation
+    and/or other materials provided with the distribution.
+
+  - Neither the name of the ORGANIZATION nor the names of its contributors may
+    be used to endorse or promote products derived from this software without
+    specific prior written permission.
+
+ THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+ AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+ IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+ DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE
+ FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
+ DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
+ SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
+ CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
+ OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE
+ USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ ====================================================================
+
+ Author:  Ikasan Development Team
+
+-->
+
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
 	<modelVersion>4.0.0</modelVersion>
 
@@ -97,6 +117,12 @@
 		</dependency>
 
 		<dependency>
+			<groupId>org.apache.activemq</groupId>
+			<artifactId>activemq-all</artifactId>
+			<scope>provided</scope>
+		</dependency>
+
+		<dependency>
 			<groupId>org.ikasan</groupId>
 			<artifactId>ikasan-scheduler</artifactId>
 			<scope>provided</scope>
@@ -137,6 +163,8 @@
 			<groupId>org.slf4j</groupId>
 			<artifactId>slf4j-simple</artifactId>
 		</dependency>
+
+
 
 	</dependencies>
